requirements.txt
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
Flask==1.1.2
gunicorn==20.0.4
heroku==0.1.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
PyJWT==1.7.1
python-dateutil==1.5
python-dotenv==0.15.0
pytz==2020.4
requests==2.25.0
six==1.15.0
twilio==6.48.0
urllib3==1.26.2
Werkzeug==1.0.1
zoomus==1.1.3
